pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "issue-tracker-api"
version = "1.0.0"
description = "Issue Tracker FastAPI Application"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-jose[cryptography]>=3.3.0", 
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "bcrypt>=4.1.2",
    "PyJWT>=2.8.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]